LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/data-url-expected.txt

CONSOLE MESSAGE: line 1: 1
CONSOLE MESSAGE: line 1: 2
CONSOLE MESSAGE: line 1: 3

PASS Test data URL and scripts errors 
